back/pyproject.toml
[tool.poetry]
name = "fastapi-frontend-auth-example"
version = "0.1.0"
description = ""
authors = ["Alexis Tacnet <alexistacnet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.62.0"
httpx = "^0.16.1"
uvicorn = "^0.13.2"
PyJWT = "^1.7.1"
SQLAlchemy = "^1.3.22"
psycopg2-binary = "^2.8.6"
pydantic = {extras = ["dotenv"], version = "^1.7.3"}
alembic = "^1.4.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
